% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzee.R
\name{auth}
\alias{auth}
\title{Authenticate with Fuzee}
\usage{
auth(username, password)
}
\arguments{
\item{username}{String - Email used to log into fuzee}

\item{password}{String -  md5 hash of password, use openssl::md5 to generate hash}
}
\value{
HTTP cookie used for GET requests downstream
}
\description{
POST request sent to Fuzee for authentication.
Password must be hashed with md5.
}
\examples{
cookie = auth(username = 'moe.dong@email.com', 
              password = 'd676b8d41fa96fd07bebe8c987513ab8')
}
