% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzee.R
\name{export_dataset}
\alias{export_dataset}
\title{Export Datasheet}
\usage{
export_dataset(dataset, cookie)
}
\arguments{
\item{dataset}{String - Name of the sheet data to download. use \code{codes()} to get a complete list}

\item{cookie}{String - cookie for authentication returned by \code{auth}}
}
\value{
a dataframe
}
\description{
Export a specific datasheet into a dataframe
A GET request is sent to fuzee for downloading a specific datasheet.
}
\examples{
batches = export_dataset(dataset = 'batches', cookie = cookie)
}
\seealso{
\code{codes()} to get a list of sheet names

\code{auth()} to authenticate with fuzee
}
