% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzee.R
\name{fuzee}
\alias{fuzee}
\title{Authenticate and Export All}
\usage{
fuzee(username, password, file)
}
\arguments{
\item{username}{String - Email used to log into fuzee}

\item{password}{String -  md5 hash of password, use openssl::md5 to generate hash}

\item{file}{String - location/file name to save zip file}
}
\value{
Returns Nothing!
}
\description{
Helper function which wraps auth() and export_all()
Use this for a one-iner
}
\examples{
fuzee(username = 'moe.doe@email.com', 
       password = 'd676b8d41fa96fd07bebe8c987513ab8',
       file = 'Desktop/all.zip')
}
