% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzee.R
\name{generate_data_url}
\alias{generate_data_url}
\title{Generate Fuzee URL}
\usage{
generate_data_url(sheet_name)
}
\arguments{
\item{sheet_name}{The name of the data sheet}
}
\value{
String - constructed URL
}
\description{
Generate URL For Data Sheets
}
\seealso{
See \code{codes()} for a list
}
