% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzee.R
\name{generate_device_id}
\alias{generate_device_id}
\title{Generate 16-bit string by random sampling letters and numbers
Used for device_id}
\usage{
generate_device_id()
}
\description{
Generate 16-bit string by random sampling letters and numbers
Used for device_id
}
