% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzee.R
\name{generate_fingerprint}
\alias{generate_fingerprint}
\title{Generate Fingerprint}
\usage{
generate_fingerprint()
}
\description{
Generate unique 32-bit string used as a fingerprint
Uses \code{openssl::md5} to generate hash
}
