% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzee.R
\name{get_codes}
\alias{get_codes}
\title{Get Sheet Names or Codes}
\usage{
get_codes(pretty = T)
}
\arguments{
\item{pretty}{boolean. TRUE pretty print sheet names, FALSE returns a list}
}
\description{
Return a list of sheet names and/or their codes
}
\details{
List of data sheets available:
Samples - Storage \cr
Batches \cr
BXA \cr
Despatch \cr
Egg Culture \cr
GA-PCRID \cr
GA-Pyro \cr
GA-Sequencing \cr
HI \cr
NAI \cr
Samples - Received \cr
Samples - Storage  \cr
Seqirus \cr
Shipments  \cr
VirusCulture  \cr
}
